Use id if it is available and fall back to key.
This happens when you are building a suggest list that is using both root fields as well as process values.

Process values store the object id and column name like "OBJECT-ID-UUID.Column Name" in the ID value.
Root fields store the field id only like "FIELD-ID-UUID" in the key value.